Fix codec correctness and robustness issues
- MelCoder: Fix run queue mask 0x3F → 0x7F to match 7-bit slot width
- HtBitIO: Use ulong arithmetic in ReadMagSgnBits to handle count=32
  (1u << 32 wraps to 1 in C#)
- HtBlockEncoder: Restore MaxPasses to 6 matching decoder capability
- CAP marker: Honour isHtOnly flag, setting bits 15-14 to HTDECLARED
  when not all blocks use HT coding
- QCD marker: Use MCT-adjusted bit depth so exponents match K_max
  computation for RGB images
- PacketDecoder: Use threshold of 256 instead of int.MaxValue for zero
  bitplane tag tree decoding to prevent near-infinite loops on
  malformed input
- JpegLs conformance tests: Fix cjpls --size flag format (WxH)

+++ src/SharpDicom/Codecs/Jpeg2000/Tier1/HtBitIO.cs
@@ -182,7 +182,7 @@ namespace SharpDicom.Codecs.Jpeg2000.Tier1
                 }
             }
 
-            uint result = (uint)_msTmp & ((1u << count) - 1);
+            uint result = (uint)(_msTmp & (uint)((1UL << count) - 1));
             _msTmp >>= count;
             _msBits -= count;
             return result;
